expense-tracker/backend/requirements-dev.txt
-r requirements.txt
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
coverage==7.3.2
django-debug-toolbar==4.2.0
